KAAV-1473 move deadline visibility parsing to separate util function
